--- src/ai/backend/manager/models/rbac_models/migration/session.py
@@ -1,11 +1,0 @@
-from collections.abc import Mapping
-
-from .enums import (
-    OperationType,
-    RoleSource,
-)
-
-role_source_to_operation: Mapping[RoleSource, set[OperationType]] = {
-    RoleSource.SYSTEM: OperationType.owner_operations(),
-    RoleSource.CUSTOM: OperationType.member_operations(),
-}
